+++ src/components/TodoApp.js
@@ -1,0 +1,12 @@
+import Header from './Heading';
+import TodosLogic from './TodoFunction';
+
+const TodoApp = () => (
+  <div className="wrapper">
+    <div className="todos">
+      <Header />
+      <TodosLogic />
+    </div>
+  </div>
+);
+export default TodoApp;
