--- src/components/TodoApp.jsx
@@ -1,12 +1,0 @@
-const TodoApp = () => {
-    return (
-      <div className="wrapper">
-      <div className="todos">
-        <Header />
-        <TodosLogic />
-      </div>
-    </div>
-    );
-  };
-  export default TodoApp;
-
