+++ src/components/TodosList.js
@@ -1,0 +1,19 @@
+import TodoItem from './ItemsTodo';
+// ...
+const TodosList = ({
+  /* eslint-disable */
+  todosProps, handleChange, delTodo, setUpdate,
+}) => (
+  <ul>
+    {todosProps.map((todo) => (
+      <TodoItem
+        key={todo.id}
+        itemProp={todo}
+        handleChange={handleChange}
+        delTodo={delTodo}
+        setUpdate={setUpdate}
+      />
+    ))}
+  </ul>
+);
+export default TodosList;
