--- src/components/TodosList.jsx
@@ -1,20 +1,0 @@
-import TodoItem from './TodoItem';
-// ...
-const TodosList = ({ todosProps, handleChange, delTodo, setUpdate }) => {
-  return (
-    <ul>
-      {todosProps.map((todo) => (
-        <TodoItem
-          key={todo.id}
-          itemProp={todo}
-          handleChange={handleChange}
-          delTodo={delTodo}
-          setUpdate={setUpdate}
-        />
-      ))}
-    </ul>
-  );
-};
-export default TodosList;
-
-
