+++ src/components/style/Header.module.css
@@ -3,6 +3,7 @@
   font-weight: 600;
   margin-bottom: 1rem;
   line-height: 1em;
-  text-transform: lowercase;
   text-align: center;
+  text-transform: uppercase;
+  color: #000;
 }
