+++ src/components/style/TodoItem.module.css
@@ -23,7 +23,6 @@
   background: transparent;
   order: 2;
   outline: none;
-  outline: none;
   border: none;
   margin-left: 7px;
 }
