+++ src/components/style/style.css
@@ -25,8 +25,6 @@ body {
 
 .form-container {
   display: flex;
-  border-radius: calc(0.1 * 100px);
-  box-shadow: 0 4px 14px 0 rgba(70, 70, 70, 0.38);
   justify-content: space-around;
   background: #fff;
   font-size: 1rem;
@@ -36,7 +34,6 @@ body {
   width: 85%;
   padding-right: 5px;
   padding-left: 10px;
-  border-radius: calc(0.5 * 100px);
   font-size: 1rem;
 }
 
@@ -67,4 +64,8 @@ body {
   display: block;
 }
 
-
+.myTodo {
+  width: 70%;
+  background: #f9f9f9;
+  padding: 5px;
+}
