+++ src/index.js
@@ -5,8 +5,8 @@ import TodoApp from './main';
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <React.StrictMode>
-  <TodoApp />
-</React.StrictMode>
+    <TodoApp />
+  </React.StrictMode>,
 );
 
 // If you want to start measuring performance in your app, pass a function
