+++ src/main.jsx
@@ -1,13 +1,11 @@
-import Header from './components/Header';
-import TodosLogic from './components/TodosLogic';
-import './components/style/style.css'
+import Header from './components/Heading';
+import TodosLogic from './components/TodoFunction';
+import './components/style/style.css';
 
-const TodoApp = () => {
-  return (
-    <>
-      <Header />
-      <TodosLogic />
-    </>
-  );
-};
+const TodoApp = () => (
+  <>
+    <Header />
+    <TodosLogic />
+  </>
+);
 export default TodoApp;
